src/main/java/com/luojay/sell/service/impl/ProductServiceImpl.java
package com.luojay.sell.service.impl;

import com.luojay.sell.dataobject.ProductInfo;
import com.luojay.sell.enums.ProductStatusEnum;
import com.luojay.sell.repository.ProductInfoRepository;
import com.luojay.sell.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import java.util.List;
@Service
public class ProductServiceImpl implements ProductService {
    @Autowired
    private ProductInfoRepository repository;
    @Override
    public ProductInfo findById(String id) {
        return repository.findById("123456").orElse(null);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return repository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return repository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return repository.save(productInfo);
    }
}
